TASK: Support language configurations with uppercase letters and underscore
When providing multiple variants of the same language, the language configuration
usually reads like "en_US" or "en_UK". The regex extracting the dimension language
currently only allows lowercase letters, which leads to extracting "en" in this case
which is not a configured language dimension.

By changing the regex to [a-zA-Z_], this case is also supported.